# api/gateway/src/main/resources/static/openapi3.yaml
openapi: 3.0.1
info:
  title: Climbing-day-member-swagger API Documentation
  description: Spring REST Docs with SwaggerUI
  version: 0.0.1
servers:
- url: http://localhost:8080
tags: []
paths:
  /v1/member/login:
    post:
      tags:
      - 회원 API
      summary: 로그인
      description: 로그인
      operationId: '로그인 API - '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1-member-login965174184'
            examples:
              '로그인 API - 실패: 로그인 정보':
                value: "{\"email\":\"test@naver.com\",\"password\":\"1234567\"}"
              로그인 API - 성공:
                value: "{\"email\":\"test@naver.com\",\"password\":\"123456\"}"
              '로그인 API - 실패: 필드 유효성':
                value: "{\"email\":\"testnaver.com\",\"password\":\"12345\"}"
      responses:
        "401":
          description: "401"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-946372771'
              examples:
                '로그인 API - 실패: 로그인 정보':
                  value: "{\"errorCode\":401,\"errorMessage\":\"아이디 또는 비밀번호를 확인해주세\
                    요.\"}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-login1698192666'
              examples:
                로그인 API - 성공:
                  value: "{\"code\":200,\"message\":\"정상 처리되었습니다.\",\"data\":{\"accessToken\"\
                    :\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoidGVzdEBuYXZlci5jb20iLCJpYXQiOjE3MTc2NzQzNDgsImV4cCI6MTcxNzY3NjE0OCwicm9sZXMiOiJST0xFX1VTRVIifQ.zQgCynIgHuMzRoJOhvSM8crsqd8iBz7M5pC8N82UzkXvPpcTcOguTJUrYvre1uWIwHP2yqdvFu-0CmqX1gIuLQ\"\
                    ,\"refreshToken\":\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoidGVzdEBuYXZlci5jb20iLCJpYXQiOjE3MTc2NzQzNDgsImV4cCI6MTcxNzczNDM0OCwicm9sZXMiOiJST0xFX1VTRVIifQ.WmKeIfehEijGe70bpODUL6L4ZaHQ3BjbKLKDvIizRy_Q0zfLB4HE1HDMBXZea7pwFoKOOiWM2mkSbduLCuxiAA\"\
                    }}"
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-1534697879'
              examples:
                '로그인 API - 실패: 필드 유효성':
                  value: "{\"errorCode\":400,\"errorMessage\":\"입력값에 대한 검증에 실패했습니다\
                    .\",\"validation\":{\"password\":\"비밀번호는 최소 6자 이상 입력해주세요.\",\"\
                    email\":\"유효하지 않은 이메일 형식입니다.\"}}"
  /v1/member/register:
    post:
      tags:
      - 회원 API
      summary: 회원 가입
      description: 회원 가입
      operationId: '회원 API - '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1-member-register389488942'
            examples:
              '회원 API - 실패: 필드 유효성':
                value: "{\"email\":\"testnaver.com\",\"name\":\"\",\"password\":\"\
                  12345\",\"passwordConfirm\":\"12345\",\"phoneNumber\":\"0101234-5678\"\
                  ,\"birthDate\":\"200011-11\"}"
              '회원 API - 실패: 이메일 인증 x':
                value: "{\"email\":\"seongo0521@gmail.com\",\"name\":\"test\",\"password\"\
                  :\"123456\",\"passwordConfirm\":\"123456\",\"phoneNumber\":\"010-1234-5679\"\
                  ,\"birthDate\":\"2000-11-11\"}"
              회원 API - 성공:
                value: "{\"email\":\"test@naver.com\",\"name\":\"test\",\"password\"\
                  :\"123456\",\"passwordConfirm\":\"123456\",\"phoneNumber\":\"010-1234-5678\"\
                  ,\"birthDate\":\"2000-11-11\"}"
              '회원 API - 실패: 중복 회원':
                value: "{\"email\":\"test@naver.com\",\"name\":\"test\",\"password\"\
                  :\"123456\",\"passwordConfirm\":\"123456\",\"phoneNumber\":\"010-1234-5678\"\
                  ,\"birthDate\":\"2000-11-11\"}"
      responses:
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-1534697879'
              examples:
                '회원 API - 실패: 필드 유효성':
                  value: "{\"errorCode\":400,\"errorMessage\":\"입력값에 대한 검증에 실패했습니다\
                    .\",\"validation\":{\"password\":\"비밀번호는 최소 6자 이상 입력해주세요.\",\"\
                    passwordConfirm\":\"비밀번호 확인은 최소 6자 이상 입력해주세요.\",\"name\":\"이름은\
                    \ 필수 항목입니다.\",\"birthDate\":\"생년월일을 올바르게 입력해주세요.\",\"email\":\"\
                    유효하지 않은 이메일 형식입니다.\"}}"
                '회원 API - 실패: 이메일 인증 x':
                  value: "{\"errorCode\":400,\"errorMessage\":\"이메일 인증이 되지 않았습니다.\
                    \ 이메일 인증 후 다시 시도해 주세요.\"}"
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register1233141654'
              examples:
                회원 API - 성공:
                  value: "{\"code\":201,\"message\":\"정상적으로 생성되었습니다.\",\"data\":{\"\
                    id\":1}}"
        "409":
          description: "409"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-946372771'
              examples:
                '회원 API - 실패: 중복 회원':
                  value: "{\"errorCode\":409,\"errorMessage\":\"이미 존재하는 회원 이메일입니다\
                    .\"}"
  /v1/member/email/auth/confirm:
    post:
      tags:
      - 회원 API
      summary: 이메일 인증 확인
      description: 이메일 인증 확인
      operationId: '이메일 인증 확인 API - '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1-member-email-auth-confirm-278006429'
            examples:
              '이메일 인증 확인 API - 실패: 인증 코드 불일치':
                value: "{\"email\":\"seongo0521@gmail.com\",\"authCode\":\"111111\"\
                  }"
              '이메일 인증 확인 API - 실패: 인증 코드 요청 필요':
                value: "{\"email\":\"seongo0521@gmail.com\",\"authCode\":\"111111\"\
                  }"
              '이메일 인증 확인 API - 실패: 필드 유효성':
                value: "{\"email\":\"\",\"authCode\":\"\"}"
              이메일 인증 확인 API - 성공:
                value: "{\"email\":\"seongo0521@gmail.com\",\"authCode\":\"111111\"\
                  }"
      responses:
        "401":
          description: "401"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-946372771'
              examples:
                '이메일 인증 확인 API - 실패: 인증 코드 불일치':
                  value: "{\"errorCode\":401,\"errorMessage\":\"인증 코드가 일치하지 않습니다.\"\
                    }"
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-1534697879'
              examples:
                '이메일 인증 확인 API - 실패: 인증 코드 요청 필요':
                  value: "{\"errorCode\":400,\"errorMessage\":\"이메일 인증 정보가 확인되지 않습\
                    니다. 확인 후 다시 시도해주세요.\"}"
                '이메일 인증 확인 API - 실패: 필드 유효성':
                  value: "{\"errorCode\":400,\"errorMessage\":\"입력값에 대한 검증에 실패했습니다\
                    .\",\"validation\":{\"authCode\":\"인증 코드는 필수 항목입니다.\",\"email\"\
                    :\"이메일은 필수 항목입니다.\"}}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-email-auth-confirm1575598845'
              examples:
                이메일 인증 확인 API - 성공:
                  value: "{\"code\":200,\"message\":\"이메일 인증이 정상적으로 처리되었습니다.\"}"
  /v1/member/email/auth/request:
    post:
      tags:
      - 회원 API
      summary: 이메일 인증 코드 요청
      description: 이메일 인증 코드 요청
      operationId: '이메일 인증 코드 요청 API - '
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1-member-email-auth-request371889986'
            examples:
              '이메일 인증 코드 요청 API - 실패: 필드 유효성':
                value: "{\"email\":\"\"}"
              이메일 인증 코드 요청 API - 성공:
                value: "{\"email\":\"seongo0521@gmail.com\"}"
              '이메일 인증 코드 요청 API - 실패: 이메일 서버 이슈':
                value: "{\"email\":\"seongo0521@gmail.com\"}"
              '이메일 인증 코드 요청 API - 실패: 이미 등록 된 이메일':
                value: "{\"email\":\"test@naver.com\"}"
      responses:
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-1534697879'
              examples:
                '이메일 인증 코드 요청 API - 실패: 필드 유효성':
                  value: "{\"errorCode\":400,\"errorMessage\":\"입력값에 대한 검증에 실패했습니다\
                    .\",\"validation\":{\"email\":\"유효하지 않은 이메일 형식입니다.\"}}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-email-auth-confirm1575598845'
              examples:
                이메일 인증 코드 요청 API - 성공:
                  value: "{\"code\":200,\"message\":\"인증 메일이 정상적으로 발송되었습니다.\"}"
        "504":
          description: "504"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-946372771'
              examples:
                '이메일 인증 코드 요청 API - 실패: 이메일 서버 이슈':
                  value: "{\"errorCode\":504,\"errorMessage\":\"이메일 전송에 실패했습니다.\"}"
        "409":
          description: "409"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-member-register-946372771'
              examples:
                '이메일 인증 코드 요청 API - 실패: 이미 등록 된 이메일':
                  value: "{\"errorCode\":409,\"errorMessage\":\"이미 존재하는 회원 이메일입니다\
                    .\"}"
components:
  schemas:
    v1-member-login1698192666:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: 액세스 토큰
            refreshToken:
              type: string
              description: 리프래쉬 토큰
        message:
          type: string
          description: 상태 메시지
    v1-member-email-auth-confirm1575598845:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        message:
          type: string
          description: 상태 메시지
    v1-member-register1233141654:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          properties:
            id:
              type: number
              description: 아이디 고유번호
        message:
          type: string
          description: 상태 메시지
    v1-member-email-auth-request371889986:
      type: object
      properties:
        email:
          type: string
          description: 이메일(계정)
    v1-member-register389488942:
      type: object
      properties:
        password:
          type: string
          description: 패스워드
        phoneNumber:
          type: string
          description: 핸드폰 번호
        passwordConfirm:
          type: string
          description: 패스워드 확인
        name:
          type: string
          description: 이름
        birthDate:
          type: string
          description: 생년월일
        email:
          type: string
          description: 이메일(계정)
    v1-member-register-946372771:
      type: object
      properties:
        errorMessage:
          type: string
          description: 상태 메시지
        errorCode:
          type: number
          description: 상태 코드
    v1-member-email-auth-confirm-278006429:
      type: object
      properties:
        authCode:
          type: string
          description: 인증 코드
        email:
          type: string
          description: 이메일(계정)
    v1-member-register-1534697879:
      type: object
      properties:
        errorMessage:
          type: string
          description: 상태 메시지
        errorCode:
          type: number
          description: 상태 코드
        validation:
          type: object
          description: 유효하지 않은 필드
    v1-member-login965174184:
      type: object
      properties:
        password:
          type: string
          description: 패스워드
        email:
          type: string
          description: 이메일(계정)